fix space advancing too far in results view
When hitting space, the UI logic would mistakenly assume the entire usable rows
were being used. This was not true when the terminal size was large enough to
have more usable rows than the number of hot keys. The result of this would be
skipping some results as you advanced between pages.