fix race conditions when checking invariants on a string queue’s head
The `check_invariant` function is called from `str_queue_pop` which is intended
to be thread-safe. So multiple threads could enter this function and then have
their unsynchronised reads of `head` in `check_invariant` race with writes
within `str_queue_pop`. This race was benign, in that the invariant will still
hold even if the head pointer changes during invariant checking. But it seems
wise to avoid this race for simplicity.