use head and tail indices instead of pointers in the string queue
This is a significant simplification that allows us to (1) not adjust head and
tail when reallocating the underlying storage and (2) drop guards/workarounds
for performing UB addition on null pointers when capacity is 0.

Note that no modulo arithmetic is required because of the two-phase assumption
of the queue. That is, because head and tail advances are never interleaved, the
ordering between head and tail never inverts.