fix: detect calls to __atomic built-ins with libclang
When parsing with libclang (`clink_parse_with_clang`), calls to the GCC __atomic
built-ins¹ were not detected. These were surfaced by libclang as unexposed
expressions. This change teaches libclink to recognise this situation and
compensate for it.

Note that this does not fix a related problem with an inability to recognise
calls to the GCC __sync built-ins.² This appear correctly as call expressions in
the AST, but their text appears as their underlying implementations. E.g.
`__sync_fetch_and_add` called on an `int` appears in the AST as
“__sync_fetch_and_add_4”.

Github: #179 “parsing C in Clang mode does not detect __atomic built-ins”

¹ https://gcc.gnu.org/onlinedocs/gcc/_005f_005fatomic-Builtins.html
² https://gcc.gnu.org/onlinedocs/gcc/_005f_005fsync-Builtins.html