reject opening any database whose schema does not exactly match our own
This change gives us a more robust versioning mechanism. The database schema can
now be changed at will, without risking users operating on an older database
with a newer Clink, causing corruption.

It might be possible to attempt some backwards compatibility. But the complexity
does not seem worth it. When the database schema changes, users are expected to
recreate any databases they have.