early-fail unit tests
The unit tests were written in a style as if the test itself is expected to
close its database connection under all circumstances (success or failure). This
is not true; `test_tmpnam` will clean up after itself even in the event of
test failure. By rewriting the tests to exit as soon as a failure is detected we
can (1) avoid some error prone coding style with variable reuse and (2)
abbreviate the amount of test code.