make 'symbols.path' a foreign key instead of direct path
This change takes a step towards normalising the database. The `path` field of
the `symbols` table is now a foreign key pointer into the `records` table. This
has a number of implications:

  1. Entries in the `symbols` table require a corresponding entry in the
     `records` table. Without this, they are effectively orphaned without a
     path.

  2. The `records` table gains a new field, `id`. Due to the way SQLite works,
     this is an alias for the internal `rowid` column. This column should rarely
     need to be referenced beyond in `join` expressions with `symbols`. Though
     `clink_db_add_record` gives the caller access to this field in anticipation
     of a more efficient API for symbol addition in future.

This change has a large effect on database on-disk size, reducing an index of a
recent Graphviz commit from 58MB to 20MB (65%).

This was not specifically intended to accelerate database construction (the
hinted at more efficient symbol addition above is the aimed for optimisation).
However this also has a large effect on this too, reducing the creation of the
above index from 7.94s to 1.81s (77%).