clink-repl: derive context on-demand when it is missing
When a file has not yet been syntax highlighted, querying the database for
results that lie within it return no context. This would go on to cause Cscope
results in Vim to appear with missing context.

This change makes the REPL derive this data on-demand when it detects it is
missing. This means that from the user’s perspective, the content is always
available in the Cscope results pane.

Github: closes #207 “lack of context in clink-repl”