clink_parse_with_clang: fix: recurse through non-main-file items
It turns out that something from the main file being parsed can have an
intermediate parent not from the main file while having a greater grandparent
from the main file:

  foo.c:42: symbol bar
   │
   └…
     │
     └ baz.h:35: symbol qux
       │
       └…
         │
         └ foo.c:45: symbol quux

When parsing something like this, we would cut off traversal as soon as we saw
`qux`, thus never encountering `quux`. I still do not quite understand why this
occurs but scenarios like the `disppath` calls in test/disppath.c have this
property. Perhaps there is a more efficient way to handle this, but for now we
just accept the cost of performing full traversal through non-main-file items.
I say this because many of these subtrees that are traversed will contain _only_
non-main-file items and turn out to be a waste of our time, but we cannot
determine this in advance.

Github: fixes #183 “some calls to 'dispath' unseen”