suppress duplicate results based only on visible fields
The name of a symbol and its column location are not visible in UI search
results. So when searching for something with multiple similar matches, results
could appear as if they were duplicates of the adjacent rows. We now only show
the first of such visually identical results.

Github: closes #205 “duplicate result rows when symbol appears multiple times”