require input paths to be absolute, when destined for the database
As we move towards storing database-relative paths, an ambiguity arises: should
relative paths passed in via the API be considered relative to the database or
relative to the current working directory? We can side step this issue entirely
by requiring them to be absolute, which already matches how clink interacts with
libclink.

Github: #15 “relative paths in database”