support F5 for rescanning files and updating the database
This is useful if you have jumped into file(s), edited them, and now want to
view updated results without exiting and restarting Clink. It does not refresh
the currently displayed search. It is not clear this is desirable. If you have a
result selected that then no longer matches after updating the database, what
should be selected after refresh?