CI: force consistent version of Python on macOS
CMake has this counter-intuitive behaviour where it will look at _all_ the
`python3` binaries in your `$PATH` and pick the one with the highest version,
instead of picking the one first in `$PATH`. The result of this is that steps
you have done to configure your Python environment beforehand (in our case
`pip3 install …`) are wasted at runtime when CMake chooses to use a different
Python.

This change attempts to force CMake to use the first binary on `$PATH`.