fix: avoid leaking /dev/null file descriptor for cscope interaction
Similar to the issue discussed in prior commits, another thread could be
attempting `fork`+`exec` causing their child to accidentally inherit copies of
these file descriptors. Problematic effects are negligible in this case, but it
is still unwise to pollute their descriptor space.

This change also has the side effect of making the child (cscope) have a
non-writeable stdin and non-readable stdout/stderr, which is slightly more
faithful.