fix: set close-on-exec on pipes to the compiler
This code is called from a multithreaded context, so it is possible for another
thread to be attempting `fork`+`exec` concurrently. If this happens, the child
produced by this `exec` can incorrectly inherit copies of these file
descriptors. I came across this realisation prompted by a recent change in the
Zig compiler,¹ though the fundamental problem is documented in the Glibc `open`
man page:

  Note that the use of this flag is essential in some multithreaded programs,
  ecause using a separate `fcntl`(2) `F_SETFD` operation to set the
  `FD_CLOEXEC` flag does not suffice to avoid race conditions where one thread
  opens a file descriptor and attempts to set its close-on-exec flag using
  `fcntl`(2) at the same time as another thread does a `fork`(2) plus
  `execve`(2). Depending on the order of execution, the race may lead to the
  file descriptor returned by `open`() being unintentionally leaked to the
  program executed by the child process created by `fork`(2). (This kind of race
  is in principle possible for any system call that creates a file descriptor
  whose close-on-exec flag should be set, and various other Linux system calls
  provide an equivalent of the `O_CLOEXEC` flag to deal with this problem.)

Unfortunately there seems to be no way to avoid this race on macOS.

¹ https://github.com/ziglang/zig/commit/389181f6be8810b5cd432e236a962229257a5b59