fix: avoid leaking /dev/null file descriptor for compiler interaction
Similar to the issue discussed in the prior commit, another thread could be
attempting `fork`+`exec` causing their child to accidentally inherit copies of
these file descriptors. Problematic effects are negligible in this case, but it
is still unwise to pollute their descriptor space.

This change also has the side effect of making the child (the compiler) have a
non-writeable stdin and non-readable stdout, which is slightly more faithful.