replace relative paths with absolute ones in compiler commands
This addresses a further aspect of compile commands database handling. Relative
paths in `-I…` and friends need to be made absolute or Clang will not find them
when we are running with a different working directory. Thus far I have been
unable to construct a test case where this makes a difference to the resulting
Clink database (hence why the test cases grep Clink debug messages), but I think
such a thing exists.

Github: #243