include database fields for semantic range
When using an external tool to consume Clink’s database, it is sometimes useful
to have this extra information available.

A minor wrinkle in implementing this is that it turns out
`clang_getCursorExtent` gives you a range that typically both starts before the
symbol name and ends after the symbol name. This is documented and even
technically correct, but just not the definition of “start” and “end” that we
use. Fortunately this matches the semantics of what the user who requested this
wants.

Github: #253
Reported-by: eaubin