make Cscope create a compressed database instead of uncompressed
Previously we were asking Cscope to make an uncompressed database because this
is simpler for us to parse. However it was observed that in this mode Cscope
enters an infinite loop when processing certain files. I attempted to report
this upstream in mail to the Sourceforge-listed contacts on 2024-12-01 but the
mail bounced. I would guess Cscope is unmaintained these days, so it is unlikely
a fix will appear any time soon. With that in mind, this change works around the
problem by using Cscope in the default mode instead, at the cost of more
complicated parsing logic.

This is a slowdown (~-9%), but I think we just have to take the hit for the sake
of avoiding this issue.

Github: closes #258 “avoid 'cscope -c'”