fix use-before-initialisation during Cscope parsing
12e8e3113855a98cf9dd32bd2c6ca4688ca2f4fa introduced a bug in this code wherein
early failures would jump past the initialisation of `pending` and
`pending_size` to code that then used these variables.

Frustratingly there seems to be no GCC or Clang warning to catch this (IMHO,
clear) mistake. -Wjump-misses-init notices it, but has so many false positives
it is not practical to turn on by default.