fix erroring out when Cscope-parsing a symlink
When parsing one or more directories of sources and a symlink was encountered,
database construction would error out. The root cause was that Clink passed the
symlink to Cscope as-is and Cscope refuses to parse symlinks. This behaviour of
Cscope appears to be deliberate, so lets just work around it. When parsing a
directory that contains symlinks, including the symlinked files seems the
desirable behaviour.