add explanatory comments to database tables
These are visible when viewed with `.schema foo` in `sqlite3`, for curious Clink
developers/users. Apart from giving people insights into the internal design,
this has the effect of a version break which will be useful in the context of
adding assignment search. Once we begin adding assignment records to the
database (upcoming commits), older versions of Clink will fail to understand
these records so it is useful to be able to force them to reject opening such a
future database.

Github: https://github.com/Smattr/clink/issues/231#issuecomment-1951450408