fix: avoid crashes when encountering '#pragma clang __debug parser_crash'
This only handles the case when the user provided no compile_commands.json or
manual Clang command line options. In these other cases, we assume the user
knows their sources will not contain this directive or they have
`-Xclang=-disable-pragma-debug-crash` in their own flags.

I do not fully understand why, but this flag does not seem to work on (at least)
LLVM ≤ 15. I have 18 locally, and we have 14, 15, and 19 in CI. It works on 18
and 19, but fails on 14 and 15. So empirically there is something in the range
[16, 18] that made this work. Dredging the LLVM Git history has turned up no
recent changes to this flag. So I have written the XFAIL as if anything ≥ 16
passes, but do not actually know if this works with 16 and 17.