turn libclang survivable crashes into a warning
This seems like a reasonable approach to an unknown outstanding set of LLVM
bugs. This required a bit of finessing because ASan somehow turns some of
these crashing scenarios into infinite loops.

This does not recover from every libclang crash. There are further uncaught
crashes we cannot easily deal with such as failing an assertion or stack
overflow.