fix: include progress bar and anticipate progress!=0 after warning
When building the database and encountering a situation in which a warning
needed to be printed, we would print the progress with the percentage as “0.00%”
and the progress bar empty. It is unclear to me why this decision was made;
progress can be (and usually is) non-zero in this scenario. This likely went
unnoticed as there is usually more progress quickly after this that then would
overwrite the progress line with a corrected version.