fix progress output creeping downwards
1fa7eb65eeeb3d13dbced6c47b6c8e51dc9443a9 introduced a bug wherein the progress
output that was meant to remain fixed at a terminal offset would creep down the
terminal. This was not typically noticeable unless starting Clink in a fresh
terminal, with a lot of vertical space below. The problem is that '\033[0B'
actually moves down a line instead of leaving the cursor where it is. The
simplest solution seems to be to omit the trailing '\n', ensuring that we are
then always moving down ≥1 when moving back to the bottom.