avoid UB in 0-byte allocator edge case
When allocating 0 bytes from an arena allocator that has not yet been
initialised, the expression `pool(me->current) + me->remaining - size` would do
arithmetic on the null pointer in a way that is Undefined Behavior. In our
usage, this (0-byte allocation calls) never happens. But it seems wise to be
robust against future changes.