fix: work around strict aliasing issues in arena allocation
Technically dealing with the unallocated memory as a `unsigned char` array and
then handing it out as arbitrarily typed pointers was UB, as I realised from a
comment¹ linked on a blog post.² I think our usage here was fine because from
this allocator we only ever allocate strings, `char *`s, that are then allowed
to alias anything. But it seems wise to make this less error prone for future
modification.

¹ https://www.reddit.com/r/C_Programming/comments/2cdpmd/comment/cjf3qhn
² https://tavianator.com/2014/alloca.html