fix: use barriers when dealing with regex list
Previously, the following sequence of actions could occur:
  1. Thread A load from `re` (re_add:38)
  2. Thread A cmpxchg to `re` (re_add:41)
  3. Thread B load from `re` (re_find:13)
  4. Thread B load from `&r->next` (re_find:17)

At this point, writes to the fields of the new entry A added to the regex list
may not be visible to B. That is, B’s `strcmp`, use of the compiled regex, or
even reads of the `.next` field may see uninitialized memory or torn values.

With `re_free`, this change is more of a paranoia step. It is assumed that
thread joining presents a memory barrier itself.