add a wrapper for running Vim from the command line
When using Clink but then running Vim outside of Clink, an annoyance is that
your Clink database is not connected. No `cscopeprg` or `cs add` adjustment has
been done, so Vim remains unaware that Clink is even in use. You can run these
commands manually to connect your Clink database, but doing this is quite
awkward.

This script is an attempt to make it more seamless to open Vim with your Clink
database pre-connected. It has only been lightly tested, so I expect this will
need to be tweaked significantly going forwards.

Reported-by: Karim Manaouil