remove 16376-byte limit on arena allocator
The arena allocator used during Cscope parsing was designed around a hard limit
of 16376 bytes. The motivation for this was an assumption that real world code
does not use symbols this large. An unexpected side effect of this was that
Cscope parsing of large strings would exceed this limit and fail. It seems
Cscope encodes strings in its database in a way that we end up needing to
decompress them to recognize and then discard them.

This change simply removes the limit, allowing allocating arbitrary sized
blocks.

Github: closes #279 “large data exceeds arena allocator’s constraints”